[[ WidgetTouch ]] Add support for touch events in widgets
This patch implements OnTouch(Start|Finish|Move|Cancel) in widgets.

The widget which hit-tests for the first touch in a sequence grabs
all touches in that sequence. The event handlers take no arguments, but instead the
full touch state is available through syntax. If the widget
becomes invisible, or is closed whilst active touches are present
Cancel is sent for all of them.

The currently active touch (the one which caused the event to
bubble) can be found by using ```the touch index```, and its
position using ```the touch position```.

The number of touches currently in progress can be found by using
```the number of touches``` and the position of a given touch using
```the position of touch tIndex```.

Note: Due to the way touch messages are currently processed above
the widget event system, widgets will receive mouse messages and
*then* touch messages. To make a widget which works on both Desktop
and Mobile, and wants touches on Mobile, the widget must have both
mouse and touch handlers, and check for platform in the mouse handlers.