[[ WidgetTouch ]] Change touch ID to sequence
This patch changes touch ID to be a sequence incremented
for each touch in a set of multi-touch events. The sequence
is reset when there are no longer any active touches so that
the first touch is always id 1 and each touch increments.

This patch also implements `the touch ids` which returns a list
of currently active touch IDs.