[[Bug 19206]] Add tests
Update the `TestClipboardFullClipboardDataAutogeneration` test to account for the change, namely that when plain text is placed on the clipboard that the other formats are not automatically generated and added.  However, any other format can be retrieved on demand and it is generated on the fly.  HTMLText is checked explicitly.

Add `TestClipboardTextOnlyOnRawClipboard` which verifies that no keys exist for HTML representations on the raw clipboard.