[[ WidgetTouch ]] Implement default (script) behavior for touch
This patch changes the way touch messages are handled when dealing
with widgets.

Touch messages are now only sent to a widget if it, or any of it
widget ancestors have at least one OnTouch* event handler.

If the first touch in a sequence is handled by a widget, then all
subsequent touches in that sequence will be grabbed by the widget
event manager.

If the first touch in a sequence is not handled by a widget, then
all subsequent touches in that sequence will be grabbed by the
engine.