[[ Bug 20565 ]] Fix linux fullscreen window manager issues
This patch removes the setting of `gdk_window_set_override_redirect`
from Linux fullscreen windows. This setting was causing the window
manager to not present other windows appropriately over the fullscreen
window. The docs note this setting should only be used for short lived
windows. This patch has the effect on ubuntu that the fullscreen window
is behind the launcher, however, this is far better than modal dialogs
being behind the fullscreen window. We are still using
`gdk_window_fullscreen` and our fullscreen property docs are very similar
to the documented behavior of that.