[[ Bug 19536 ]] Fix menu not closing when clicking on other side
This patch  fixes a bugh where the stack of the button with the menu
was being used to determine if the user clicked away from the menu
rather than the menu stack itself. This was causing the menu to only
dismiss if the use clicked in a location less than the top left of the
button.