[[ Bug 20961 ]] Reset the menu object when swapping menus
This patch resets `MCmenuobjectptr` to `null` when changing from one
menu to another. This allows `openmenu` to set it to the new button and
and for `closemenu` to correctly set it to `null`. Previously
`MCmenuobjectptr` was left referring to the first button and therefore
not cleared on `closemenu` and leaving the button in an unable to delete
state.