[[ Bug 21109 ]] Limit clipboard clear setting fullClipboardData["text"]
This patch ensures the clipboard is cleared only when necessary when setting
`the fullClipboardData["text"]`. It also ensures that private data is never
cleared. Additionally the clearing of the clipboard when there are styled
text representations and the text representation is set is now documented
as part of the interaction between the styled and plain text representations
placed on the clipboard.

(cherry picked from commit 90e1cd9a2ca5f2ef1e4f328feac31d49b13a134b)