[[ Bug 20488 ]] Ensure put throws when expecting preposition
This patch ensures that the put command throws an error if it parses
a token when expecting a preposition and it isn't one. Previously
the command backed up and returned normal resulting in the anomaly that
anything after `put <expression>` could be treated as a second statement
without a newline or `;` present.