[[ Bug 21099 ]] Fix crash closing browser
A previous patch allowed CEF to complete browser closing by returning
`false` in the `DoClose` callback. This allowed CEF to cleanup the render
process for the browser. The patch should have also removed the browser
window destruction allowing CEF to do that. This caused a crash on some
versions of Windows.