[[ Bug 20488 ]] Fix docs and test
Prevously the `catch` test parsed because the test placed `put` before
the example. The catch example needs to be skipped like `case` and `end`.

The try control structure example was taking advantage of the lack of parser
error and expecting an execution error for `in` not being a command.

The `revBrowserSet` example clearly shows why we need a parser error here.