[[ Bug 19753 ]] Prevent chunk out of range error on Linux when clicking selection
At times on Linux, the `mHoverRow` variable can get set to 0 and cause clicks on the widget to generate a "chunk out of range" error.  Added a check to the top of the `OnClick` handler to detect this condition and set the value if needed.  This solves an additional issue where if the widget is on a window that does not have focus, then the `mHoverRow` variable would not get set until the mouse moved after the click to bring the window forward.