[[ Bug 20142 ]] Prevent index errors when expanding TreeView buffer beyond 1000 keys
The process to load additional data didn't properly account for level 0 `more_data` expansions.

`addToListBuffer` trims one level off of a path, but did not account for the first level.  Added an `if` block to return an empty list if the path only contains a single item.

`getValueOnPath` did not account for an empty list (which needs to return the whole array) which happens when the base list contains more then 1000 items.  Also moved the key validation check to before the first use of the key.

Converted spaces to tabs in a section of `calculateMoreElements` and corrected indent levels.