[[ Bug 19754 ]] Enhance TreeView numeric sort to sort non-numbers
Updated `CompareKeysNumeric` handler so that if both values are non-numbers, then it will sort them instead of returning 0.  Completely removed the 0 return value since keys may not be duplicates.  In the case of numbers that have the same numeric value, sort based on the text representation to have a consistent result (e.g. always have 02 come before 2).  In the case of a descending sort, all numbers remain grouped at the top of the list.