[[ Bug 21203 ]] Only post hiliteChanged when value actually changes
Switch Button widget posted the `hiliteChanged` message whenever a click was registered within the rect of the widget.  It also would post the message in other instances where the value didn't actually change.

Added code to detect if the value was being changed and only issue the message when the value actually changes.