[[ Navigation Bar ]] Add option to highlight none of the icons
Update documentation references for the new 0 option to have nothing highlighted.
Add ability to set the selected icon name to "" to make nothing highlight.
Updated documentation to indicate that custom icons are possible (reference Icon SVG Library).
Renamed note and expanded to document potential backward compatibility issue (a call to `getNavSelectedItemName` in a previous version of the widget with nothing selected will throw an error).