[[ Widget Native Layer Focus ]] Sync native layer focus and control focus
This patch ensures that native layer focus is synced with control focus allowing
`the focusedObject` to be correct for widgets with native layers and removing
focus from appropriately from other objects. At present `the traversalOn` of the
widget must be true in order for it to become the focused object regardless of
whether it currently has the keyboard focus.