[[ Bug 21246 ]] Fix crash on windows clipboard
The patch for Bug 20678 caused a number of regressions on the windows
clipboard. Under certain circumstances this would cause a crash, however,
it also caused a failure to ush data to the clipboard. This patch reverts
the changes in that patch and fixes Bug 20678 in a way that does not cause
those regressions.