[[ Bug 15443 ]] Don't clobber macroman casing tables on mac MCS_tolowercase and MCS_touppercase were returning incorrect values for characters in the upper half of the native char set on mac due to having the macroman casing tables clobbered by the utf-16 version.
The casing code on each platform is sufficiently different (it is
only shared by android and iOS, and linux and emscripten) that there
doesn't seem to be much point keeping it in MCS_common_init - each
platform was doing what it needed in MCSystemInterface::Initialize
anyway.

This patch also moves the macroman casing table construction outside
the #ifndef _MAC_SERVER block, as it is needed on Mac server.