[[ SVG ]] Add support for 'currentColor' paint type
This patch adds support for the SVG 'currentColor' paint type.

The backgroundColor property of an image containing a drawing is
used as the value for 'currentColor' when painting the drawing.

In implementing this feature, more progress towards cleaning up the
management of paints internally has been done. New public libgraphics
constructors for solid color, pattern and gradient paints have been
added, and these are now used by the canvas API and internally.