[[ ClipToPath ]] Implement clip to path
This patch implements a new statement in the lcb canvas library to clip
to a path. Previously only setting a clipping rectangle was supported
however this patch allows for more complext cliping paths. It additionally
simplifies the clipping of images as there is no need to create a pattern
paint from them and fill a path.