[[ Drawing ]] Fallback to relative resource path
This patch ensures that if there is no library mapping set for the resource
path in the current environment (whether IDE or not) that the fallback is to
return the path relative to the stack file. This allows the library to be
loaded in environmentes where it was not included in standalone building.