[[ Bug 21336 ]] Skip rendering if ECS_DONTDRAW true
This patch ensures that MCStack::view_render returns immediately if
`ECS_DONTDRAW` is set. This exstended state is set/unset on a stack
when a new stack is opened on mobile causing it to become the stack
which should be rendered to screen. Previously portions of other
open stacks could be rendered over the top stack.