[[ Bug 16366 ]] Fix ShowAll mode dirty rect
This patch fixes an issue where showAll fullscreen mode was not correctly dirtying
the entire rect that needed to be re-rendered causing the letterbox area to remain
the same as the previous card. The issue was twofold:

- The code paths for dirty in the entire stack were not consistent with many places
calling view_dirty_all and many calling dirtyall
- view_dirty_all was not dirtying the entire visible stack rect

This patch removes `view_dirty_all` by routing `dirtyall` through `dirtyrect` to
`view_dirty_rect`.