[[ Bug 21356 ]] Fix black screen with acceleratedRendering at startup
This patch fixes the issue where if the acceleratedRendering was set to true
in a preOpenStack or startup handler on Android the SurfaceHandler callbacks
were never called. By moving the creation and adding of the view to a runnable
this issue no longer occurs.
(imported from commit a66d49a17c0ec13ef7682301359010eb43777a29)