[[ Bug 21386 ]] Set clipsToBounds of native container layer
iOS UIViews do not clip child views to their bounds by default so the
child view is drawn iv it is on screen regardless of whether portions
are outside the bounds of the parent view. This patch sets the clipsToBounds
property of native container layers so that they will clip to the rect
of the group as intended.