[[ Bug 21434 ]] Force bitmap view visible for visual effects
This patch fixes a regresion from a previous patch that changed the behavior
of the `showBitmapView` method to only set the visibility of `m_bitmap_view` to
visible if the OpenGL view is not visible. This interfered with visual effects
as they uset the bitmap view rather than OpenGL view.