[[ Engine ]] Add 'layerClipRect' property to controls
This patch adds a new property 'layerClipRect' to all controls.

The property defines a clipping rectangle in card co-ordinates which
is applied both when the control is rendered normally, and also to
the accelerated rendering layer-clip when accelerated rendering is
turned on.

The property means that the restriction of only top-level objects
being able to have dynamic layers in accelerated rendering mode
is not quite such a problem as it is now possible to have a group
clipping like affect applied to top-level objects.