[[ Bug 10881 ]] Fix keyboard panning view when acceleratedRendering true
This patch changes the use of `SOFT_INPUT_ADJUST_RESIZE` to
`SOFT_INPUT_ADJUST_NOTHING` when the keyboard is presented. The OpenGLView was
not resizing and as a result was being pushed up by the keyboard and then
not returning. Additionally it is unnecesary to resize and therefore reconfigure
the bitmap when the keyboard is presented.