[[Feature]] Add OS and Platforms to LCB Metadata
Added new metadata keys to `generate.g` file

Added `revSBNormalizePlatforms` to `revSBLibrary`. This new handler will
take a list of OSs in the dictionary format and convert them to the
standard names that used by the SB. If no OSs are specified, then the
list of Platforms will be used. As a fall back, if the list remains
empty (due to invalid data in the OS or Platforms lists), then it will
revert to the previous behavior and supply a list of all platforms.

Add `metadata` to the `revDocsExtractDocBlocks` handler of
`revDocsParser`. This block will add the metadata keys to the library
comment block so that they get included in the dictionary. The initial
list of keys is `author, os, platforms, version` but only OS and
Platforms are in the actual dictionary (the other 2 are in the generated
lcdoc files).

Updated Element lists for `revDocsParseDocText` to inherit `os` and
`platforms` from Library.  Updated Library elements to add `os`,
`platforms`, and `associations`.  Added case statements for `os` and
`platforms` and removed trailing white space in the area.

Updated `extensionFetchMetadata` in `extension-utils` to parse the
platforms and os keys that were added.