[[Bug 21488]] Check 'uname -m' if on Linux and 'uname -p' is unknown
Fixed bug 21488, where on some Linux systems the Makefile is unable to
automatically determine the architecture to build for. This is due to
the Makefile using 'uname -p', which can return 'unknown' in some cases.

Fixed by adding an additional case statement into the Makefile that
checks 'uname -m' if 'unknown' is returned from 'uname -p'.