[[Bugfix 19619]] The outer nub inherits its color from the minute hand
The minute hand of the clock is the first hand to be drawn, directly above
the outer nub. In the same way that the inner nub (which is drawn last)
inherits the color of the second hand (which is drawn second-last),  the
outer nub should probably inherit the color of the minute hand.