[[Bugfix 19619]] Updated overview documentation and version number
The widget overview documnetation has been updated to include information
abou the new properties that allow for custom styling of the clock.

The version number of the widget has been increased by 0.1, in-line with
Semantic Versioning (the property changes are backwards compatible, older
clock widgets can be imported and will continue to function as they did
before).