[[Bugfix 19619]] Store color properties in Strings, not Colors
The ten various color properties of the clock widget are now stored in
strings rather than color variables. This simplifies the code by
elimintating the need for seperate getters for all the color properties.
Instead, the conversion from a string to a color is done whenever the
color is required for use in the OnPaint handler.