[[Bugfix 19619]] Updated test file
The test script for the properties of the clock has been updated
to reflect the new method of storing the color properties as strings.
With this new method, if the color is set to a string that does not
include an alpha value, then no alpha value will be returned.