[[ Bug 21396 ]] Fix crash on startup in iOS 12 beta
This patch fixes an issue where the view layer appears to be retaining the
`CGDataProvider` from an image when it is drawn via `CGContextDrawImage`.
As we were freeing the raster directly after the draw the `CALayer` could not
access it from the `CGDataProvider` when rendering causing a crash.

This patch therefore changes the behavior of `MCGRasterCreateCGDataProvider`
where previously when calling it with the copy parameter `false` it would be
left to the caller to free the buffer the data provider now takes ownership
of it and it is freed via the data provider free callback. As a consequence
of the change custom cursor creation on mac now copies the buffer for
simplicity.