[[ Module Lcext ]] Update iOS Standalone builder handle module lcexts
This patch updates the standalone builder to link module lcexts and include the
module.lcm for them which is just the name of the module. Most noteworthy is the
removal of deadstripping by the standalone builder. All lcexts being linked together
should be dead stripped when built. This change allows us to link lcexts without
forcing the `__libinfoptr_<name>` which is the method used to link externals and
is not in the module lcext.