[[Bugfix-14721]] Added go visible form of go command
The go command has been updated to include the go visible form
which automatically sets the visible property of the stack to
true when it is opened.

This is particularly useful for script-only stacks that display
any kind of interface (ie: that have visible controls), as they
currently need to manually include a line in the openStack or
preOpenStack handler to set the visible property to true.