[[Bugfix-14721]] Add parser and script tests for go visible
A new parser test file has been created for the go command, as it did not
have one before. Three new tests have been added, to test the basic syntax
of go stack "stack", go visible stack "stack" and go invisible stack "stack".
This test file is far from complete - there are many other forms of the go
command that have not been tested, however I have added in only the three
most relevent to the go visible command. Any others can be added separately
in a different branch.

Several new tests for the go command have been added to the core interface
test file. These tests check that the stack actually opens when using the
visible and invisible forms and that the visibility of the stack is correctly
set.