[[Bugfix-14721]] Updated dictionary for go visible
The dictionary file for the go command has been updated to include
the new go visible command. A note has also been added to point out
the occasion where the go visible and go invisible commands will
not work as expected (when the stack overrides the command by setting
its visible property in its script as it opens).