[[Bugfix 19619]] Added dependancies for clock to tests
Two of the test files required the clock module in the tests.
The new clock module (version 1.1.3) now requires the widgetutils
library, and if this was not loaded prior to loading the clock
module then the test would fail.

For some reason all these tests run fine on my machine even when
the widgetutils dependancy is not loaded prior to the clock module,
so I cannot confirm 100% that all the tests will run fine this time!