[[Docfix]] Added undocumented TestSkipIf requirements
The testing.md file which explains the process for writing tests
for the various aspects of LiveCode contains a section on the
TestSkipIf command, which skips a test if a certain requirment
is met. However not all the available requirement options that
TestSkipIf provides were documented.

Information for these requirement options has now been added.