[[Bugfix-14721]] Represent visibility type by enum
The two boolean variables that represented the visibility of the stack,
p_visible and p_explicitly_visible, have been replaced by a single enum
p_visibility_type, of type MCGoStackVisibilityType, which has three
options: kImplicit (if the regular go <stack> form is used, so the
visible property of the stack remains unchanged), kExplicitVisible (if
the go visible <stack> form is used, so the visibility of the stack
should be set to true) and kExplicitInvisible (if the go invisible <stack>
form is used, so the visibility of the stack should be set to false).

The visible variable of the MCGo class (declared on line 1740 of cmds.h)
has not been removed as it appears to be used by several other parts of
the engine.