[[Bugfix-14721]] Increase readability of enum
The readibility of the MCGoStackVisibility enum has been increased by
renaming it to MCInterfaceExecGoVisibility, and prepending the new enum
name to all the enumeration values.

Also the use of braces in the enum definition has been updated to conform
to the existing style (note to self: read C++style.md *first* next time).