[[ Bug 21503 ]] Tree widget - prevent value removal on key addition
Current operation of the tree widget will delete the value stored in a key
if the "+" is clicked to add a new sub-key and it is not already an array.

This change alters the behavior such that when the "+" is clicked, if the
current key is not already an array, then the existing value is moved to
the new sub-key that is created.