[[ Bug 21295 ]] Fix synonym handling of widget properties
This patch fixes synonym handling by finding the property name that the
widget is using rather than the first property name for a particular
property. The patch also handles unimplemented engine properties as an
error as with legacy objects.

While setting unimplemented engine properties as a custom property may
be useful it should be consistent for all objects and is not possible
if the only information is the value of the Properties enum as we would
need the name used in the script.