[[ Bug 19670 ]] Fix effective working screenrect on Android
This patch fixes the effective working screenrect on Android when
the status bar is hidden. Previously to determine the area not covered
by screen decorations we used `getGlobalVisibleRect`, however this only
took clipping into account and in some cases decorations do not clip the
view.

This patch uses `getWindowVisibleDisplayFrame` which is documented to tell
`you the available area where content can be placed and remain visible to
users.`