[[ Deploy ]] Handle payload and project out of order on windows
This patch allows deploy to handle installer standalones that have the payload
and project sections out of order. Previously it was assumed that the last three
sections of the installer were `.payload`, `.project` and `.rsrc`. This patch
allows for the section order to also be `.project`, `.payload`, `.rsrc`, however,
it is still assumed that they are the last three sections and `.rsrc` is the last
section.