[[ Bug 21623 ]] Handle stack scroll in open in window
This patch ensures that the stack scroll is taken into account when taking
a window from another stack. The stack taking the window is closed and therefore
the scroll is cleared, however, the stack with the target window is not closed
so may have a non-zero `vScroll`. As a result we need to adjust the rect with
the `vScroll` after copying it.