[[ Cleanup ]] Refactor the processor to reuse code
This patch refactors `the processor` to reuse code between platforms
and also between LCS and LCB. A new `MCSInfo` api was implemented in
libfoundation with `MCSInfoGetArchitecture` returning a string based
architecture definitions in `foundation.h`.

Additionally `the processor` documentation was updated to clarify
returned values in emulated environments and to disassociate the
returned values with particular operating systems to reduce maintenance.