[[ Bug 21624 ]] Fix blendlevel + disabled button issue for Motif
This patch fixes an issue where we begin a layer but do not end it because
the call to `dc->end()` is only called when a label is being shown for the
button. The fix is to move the call out of that code block. It is safe to
do so as the only code after it is for Win 95 appearance.