[[ Tree View Widget ]] Auto expand to show selected row
The tree view widget now has the ability to automatically expand to reveal
a row when it is selected. If **scrollHilitedElementIntoView** is not `true`
then the scroll position will be adjusted to maintain the currently visible
top row.

The tree view fold state can now be reset in two ways. To set the array
data in a folded state then use the **foldedArrayData** property. To
collapse the tree without changing the data, get the **foldedArrayData**.

Added 2 module variables to control redraw and determine when fold state
changes (`mRedraw` and `mFoldChanged`).

Added `pathIsValid` handler to ensure that a user supplied path actually
exists.  Validation is now done in the `setSelectedElement` handler so the
requirement to test for an empty path in `selectPath` is removed.

Added `unfoldFullPath` handler to take care of ensuring the full selection
path is unfolded.  Code is included to maintain the top visible row if
`mScrollHilitedElementIntoView` is not `true`.

Updated `applyFoldStateKey` handler to set `mFoldChanged` when an expansion
is performed.  Updated to check for the need to expand.