[[ Bug 21474 ]] Reset mouse stack after drag and drop
This patch resolves an issue where after a drag and drop operation the mouse stack
was being cleared in `wmdragdrop`. This resulted in a seemingly unresponsive stack
as mouse events are guarded by `MCmousestackptr`. As `wmdragdrop` is only called when
a window recieves a drag drop event we can ensure the mouse stack is correct by finding
the stack for the window and setting it to that.