[[ Bug 20659 ]] Update TreeView to ignore mouseUp when row has changed
Add `mMouseDownRow` to record the row clicked when a mouseDown event is
registered that does not take place in the scrollBar or Separator.

Add check to the `OnClick` handler to ensure that clicks are ignored if
the mouseDown and mouseUp rows are different.

Moved the `addBaseLevelElement` handler to ensure that it is not called
when mouseDown and mouseUp rows are different.

Converted spaces to tabs in several places.