[[ Bug 21623 ]] Fix early window synchronize when taking window
This patch fixes an issue when opening a stack in another stack's
window. The call to `view_setstackviewport` before the stack's opened
value is incremented was causing `view_setgeom` to ignore the screen
lock. Additionally, this patch adds a call to `dirtyall` before stack
scroll is applied. This is allows the region that is to be scrolled
to redraw correctly.