[[ TreeView ]] Get/Set Fold State
Add `foldState` property to allow the full fold state to be retrieved
and managed via LCS.  This will allow arbitrary states to be saved and
restored along with the ability to clear the fold state.

Checks were added so that invalid data does not crash the widget.
If the `folded` key is missing for a key in the array, then it is skipped.
If a boolean is not set for the `folded` key value, then the value is
ignored.