[[ TreeView ]] Add hilitedElementFoldState property
Add a property to allow the fold state of the selected element to be
get and set.  This also allows determining if the selected element
is a leaf node.

Updated documentation for enum parameters so they are formatted properly.

Moved getSelectedElement handler so the ordering of get/set handlers is
more consistent (get handlers appear before set handlers).