[[ Engine ]] Add variant of params() which returns a sequence
This patch adds a variant of the params function.

The params function can now take 0, 1 or 2 parameters.

Calling it with zero parameters has the same behavior as before -
it returns a stringified version of the call that was made.

Calling it with two parameters returns a range of parameters
as a numerically sequenced array. The first parameter indicates the
starting index (from 1), and the second parameter indicates the
finishing index (from 1). Both parameters may be negative, in which
case they are interpreted as being relative to 'the paramCount' in
the same way negative chunk indicies work.

If only one parameter is specified, the second is taken to be -1.