[[ Foundation ]] Make __MCArrayIsEqualTo recursive
This patch changes the implementation of __MCArrayIsEqualTo so that
it recurses to MCValueIsEqualTo on values. This should improve the
veracity of interring an array, as it will mean that two arrays
only need to have the same values as contents, regardless of whether
those values have been interred to inter to the same value.