Test AdjustWindowRectExForDpi
This commit adds a weak-linked `AdjustWindowRectExForDpi` function with a hardcoded dpi value for testing. It confirms that using `AdjustWindowRectExForDpi` from within `getwrect` draws a window correctly on a secondary display with a different resolution than the primary display.

This test was run on a dual monitor setup. The primary displays was set to 96 dpi (1.0 pixel scale) and the secondary display was set to 144 dpi (1.5 pixel scale). The “save changes to stack” dialog is drawn at the correct height on the secondary display using this code, whereas it is drawn too short when using the current LC code.