[[ Bug 21685 ]] Fix android keyboard resize handling
Recent changes to the keyboard handling on android resulted in the instance variable
`m_keyboard_visible` being out of sync with the actual keyboard visibility. This patch
changes that variable to be purely a sentinel for `updateKeyboardVisible` and moves
the keyboard visibility detection from the global layout listener to a new `keyboardIsVisible`
method.