[[ Simple Case ]] Add locale parameter to toLower & toUpper
This patch adds a locale parameter to `toLower` and `toUpper` allowing
scripts to specify the locale that should be used for case mapping. If
empty then the simple non-language specific case mapping will be used.

The new parameter is optional and without the parameter the detected
system locale continues to be used for case mapping.