[[ Foundation ]] Minimise size of core value structs
This patch minimises the size of the libfoundatino core value
structures. Specifically it shifts the order of some fields in
__MCString and __MCData resulting in a reduction in size on
64-bit. Additionally, on 64-bit, the hash field has been removed
from __MCName with 28-bits being placed in the flag word, and
the remaining 4-bits being split between the lowest two bits
of the key and next fields - this means that hash-order is
preserved.

The effect on this on memory utilization will vary depending
on platform and architecture - most critically, the effacy of
the implementatin of malloc for small object sizes.

However, on 64-bit Mac, it means that __MCName and __MCString
will both fit into a 32-byte (no overhead) malloc block, as
opposed to a 48-byte (no overhead) malloc block as occurred
previously.