[[ Bug 21860 ]] Fix encoding of svg paths containing relative commands
This patch fixes a problem in svg paths making heavy use of relative
commands caused by close subpath not resetting the last point to
the first point of the subpath.

The first point is now recorded when a move command is encountered,
and this is used when a close command is encountered to update the
last point.