[[ Bug 21915 ]] Fix memory leak in 'the processor'
The refactoring of the internal support code the 'the processor'
function resulted in the introduction of a memory leak due to
over-retaining the return value of MCS_getprocessor().

The leak has been fixed by changing the signature of MCS_getprocessor()
to match the standard style of bool return value to indicate an error
and an out parameter to take the string result.