[[ Bug 10881 ]] Fix Android openGL compositor surface height issue
This patch fixes an issue on Android where the SurfaceView is being resized
when the keyboard is activated and deactivated. We do not resize the view for
keyboard activation and as the openGL coordinate system has bottom-left origin
the bottom of the stack was being rendered where the top should be.

This patch changes the openGL tile cache to request the surface height directly
through EGL APIs. As these APIs are unavailable on iOS the change is `#ifdef`d
for Android only.

Additionally to ensure the stack is re-drawn without necessarily re-rendering it
the patch implements a new `MCScreenDC::refresh_current_window()` API. This is
particularly important for when the keyboard is hidden as otherwise the area
previously occupied by it will be black.