[[ Bug 21918 ]] Memory leak when manipulating four-char-codes on macOS
This patch fixes a leak in the macOS platform specific FourCharCodeFromString
function. In order to convert a string into a FourCharCode integer, the
string is temporarily converted to a CString. Previously, however, this
temporary CString was not being deleted. The explicit conversion has been replaced
by an AutoStringRefAsCString auto class which fixes the leak.