[[ Bug 21919 ]] Fix memory leak when deleting system font refs on macOS
This patch fixes a leak which occurs when references to system/theme fonts
are deleted on macOS. System font handles are cached globally separately
from the higher-level management of them (specifically in the theme-related
code) so mustn't be released. However, the MCFontstruct which is created
to wrap them must still be deleted. The leak occurs when the engine must
'purge fonts' to update references after a font has been explicitly
loaded and unloaded.